[TASK-3] fixes

+++ ksenia.ivanova/task-3/cmd/service/main.go
@@ -36,7 +36,7 @@ func main() {
 		outputCurrencies[i] = model.OutputCurrency{
 			NumCode:  valute.NumCode,
 			CharCode: valute.CharCode,
-			Value:    valute.Value,
+			Value:    float64(valute.Value),
 		}
 	}
